# Conflicts:
#	ad-hell/src/main/java/com/adhd/ad_hell/domain/ad_favorite/command/domain/aggregate/AdFavorite.java

+++ ad-hell/src/main/java/com/adhd/ad_hell/domain/user/command/dto/response/UserDetailResponse.java
@@ -17,7 +17,6 @@ public class UserDetailResponse {
     private UserStatus status;
     private LocalDateTime createdAt;
     private LocalDateTime deactivatedAt;
-    private Long amount;
-
+    private Integer amount;
 
 }
